Fix issues when adding new Car
- Change few attribute types in Car (int to float)
- Fix connection between Car and Cart table
- Refactor Joi types for new_car endpoint

+++ server/src/db/models/CarC.ts
@@ -132,7 +132,7 @@ export const CarC = (sequelize: Sequelize): CarStatic => {
         allowNull: false
       },
       mileage: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.FLOAT,
         allowNull: false
       },
       engineSize: {
@@ -171,7 +171,7 @@ export const CarC = (sequelize: Sequelize): CarStatic => {
         allowNull: false
       },
       bootCapacity: {
-        type: DataTypes.SMALLINT,
+        type: DataTypes.FLOAT,
         allowNull: false
       },
       AC: {
@@ -203,7 +203,7 @@ export const CarC = (sequelize: Sequelize): CarStatic => {
         allowNull: false
       },
       price: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.FLOAT,
         allowNull: false
       },
       images: {
